Cast

+++ app/src/main/java/com/tmendes/birthdaydroid/BarChartActivity.java
@@ -108,7 +108,7 @@ public class BarChartActivity extends AppCompatActivity  implements OnChartValue
 
     @Override
     public void onValueSelected(Entry e, Highlight h) {
-        String msg = Float.toString(e.getY());
+        String msg = Integer.toString(((int) e.getY()));
         Toast.makeText(this, msg , Toast.LENGTH_SHORT).show();
     }
